build(windows): complete the native development profile
Build the CLI, TUI, and native tests from one debug or release flavor tree, with test dependencies enabled by default. Keep build state checkout-isolated on local storage, centralize Visual Studio discovery, and let command modules declare when the MSVC environment is required.

Represent audio backends through a generated platform capability header, make Linux backend dependencies explicit, and add portable filesystem access test support. Enable MSVC warnings as errors and adapt CLI reflection types, platform code, and tests so the complete native graph builds cleanly.

Harden LLVM 22 hygiene by building the Windows tidy executable warning-free, refreshing stale CMake feature state, and replaying compile commands through narrowly filtered temporary databases. Preserve MSVC flags in product builds, document the host-specific behavior, and cover the portal changes with tooling tests.

+++ test/unit/audio/detail/Mp4PacketSourceTest.cpp
@@ -23,8 +23,8 @@ namespace ao::audio::detail::test
       ao::test::mp4::addAtom(data, "mdat", payload);
 
       auto const config = ao::test::mp4::makeAtom("alac", {9, 8, 7});
-      auto const track =
-        ao::test::mp4::makeCompleteAudioTrackAtom("alac", config, timescale, 88200, payload.size(), 1024, 8);
+      auto const track = ao::test::mp4::makeCompleteAudioTrackAtom(
+        "alac", config, timescale, 88200, static_cast<std::uint32_t>(payload.size()), 1024, 8);
       auto const moov = ao::test::mp4::makeAtom("moov", track);
       data.insert(data.end(), moov.begin(), moov.end());
       return data;
